Split of ContactFinder from CollisionFinder
